renaming the folders and organizing the blocks
