+++ src/categories/javascript.ts
@@ -1,22 +1,36 @@
-import "../blocks/javascript/blocks";
-import "../blocks/javascript/general_blocks";
+import "../blocks/javascript/beginner";
+import "../blocks/javascript/form_blocks";
+import "../blocks/javascript/todo_blocks";
 
 export const javascriptCategory = {
   kind: "category",
   name: "Javascript",
-  colour: "#5CA699",
+  colour: "#f39c12",
   contents: [
     {
-      'kind': 'category',
-      'name': 'Beginner JavaScript',
-      'contents': [
+      kind: "category",
+      name: "JS Beginner",
+      colour: "#FF007F",
+      contents: [
         {
           kind:"block",
           type: "javascript"
         },
         {
           kind:"block",
-          type: "console_log" 
+          type: "create_variable"
+        },
+        {
+          kind:"block",
+          type: "generate_id" 
+        },
+        {
+          kind: "block",
+          type: "chnage_innerHTML"
+        },
+        {
+          kind: "block",
+          type: "console_log"
         },
         {
           kind: "block",
@@ -24,15 +38,7 @@ export const javascriptCategory = {
         },
         {
           kind: "block",
-          type: "create_variable"
-        },
-        {
-          kind: "block",
-          type: "event_listener"
-        },
-        {
-          kind: "block",
-          type: "generate_id",
+          type: "event_listener",
         }, 
         {
           kind: "block",
@@ -41,9 +47,30 @@ export const javascriptCategory = {
       ]
     },
     {
-      'kind': 'category',
-      'name': 'Main Functions',
-      'contents': [
+      kind: "category",
+      name: "JS Form",
+      colour: "#800080",
+      contents: [
+        {
+          kind: "block",
+          type: "handle_form_submission",
+        }, 
+        {
+          kind:"block",
+          type: "set_form_data_to"
+        }, 
+        {
+          kind: "block",
+          type: "fetch_block"
+        },
+        {
+          kind: "block",
+          type: "clear_form_fields",
+        }, 
+        {
+          kind:"block",
+          type: "auto_fill_form_fields" 
+        },
         {
           kind: "block",
          type: "validate_and_handle_error", 
@@ -64,37 +91,18 @@ export const javascriptCategory = {
          kind: "block",
          type: "show_data_in_alert_custom",
        },
-       {
-        kind: "block",
-        type: "auto_fill_form_fields",
-      }, 
       ]
     },
     {
-      'kind': 'category',
-      'name': 'Form Functions',
-      'contents': [
-        {
-          kind: "block",
-          type: "generate_form_id",
-        }, 
+      kind: "category",
+      name: "JS Todo",
+      colour: "#FFD700",
+      contents: [
         {
           kind:"block",
-          type: "handle_form_submission"
+          type: "add_todo_task"
         },
-        {
-          kind: "block",
-          type: "set_form_data_to",
-        }, 
-        {
-          kind: "block",
-          type: "fetch_block"
-        },
-        {
-          kind: "block",
-          type: "clear_form_fields",
-        }, 
       ]
-    }
+    },
   ],
 };
